backend/requirements.txt
fastapi
uvicorn 
requests
motor
pydantic
python-dotenv
google-generativeai
jsonschema
typing
PyPDF2
python-multipart